--- graphrag/index/graph/__init__.py
@@ -1,4 +1,0 @@
-# Copyright (c) 2024 Microsoft Corporation.
-# Licensed under the MIT License
-
-"""The Indexing Engine graph package root."""
