--- graphrag/index/graph/extractors/__init__.py
@@ -1,12 +1,0 @@
-# Copyright (c) 2024 Microsoft Corporation.
-# Licensed under the MIT License
-
-"""The Indexing Engine graph extractors package root."""
-
-from graphrag.index.graph.extractors.community_reports import (
-    CommunityReportsExtractor,
-)
-
-__all__ = [
-    "CommunityReportsExtractor",
-]
